+++ packages/manager/src/MainContent.tsx
@@ -117,11 +117,6 @@ const LinodesRoutes = React.lazy(() =>
     default: module.LinodesRoutes,
   }))
 );
-const Kubernetes = React.lazy(() =>
-  import('src/features/Kubernetes').then((module) => ({
-    default: module.Kubernetes,
-  }))
-);
 const Profile = React.lazy(() =>
   import('src/features/Profile/Profile').then((module) => ({
     default: module.Profile,
@@ -324,10 +319,6 @@ export const MainContent = () => {
                                 <Route
                                   component={LinodesRoutes}
                                   path="/linodes"
-                                />
-                                <Route
-                                  component={Kubernetes}
-                                  path="/kubernetes"
                                 />
                                 {isIAMEnabled && (
                                   <Route component={IAM} path="/iam" />
